Step 3: move chronological ordering into trace.py
The pipeline-specific ordering logic (build_chronological_frames and all
its helpers) now lives in trace.py alongside the other pipeline stages.
chain_analysis.py is reduced to pure AST/try-except utilities plus thin
backward-compatible wrappers so existing imports keep working.
